Revert "[refactor] moving memtracer to gemini (#790)"
This reverts commit 1477c387ad28062163f00594076383f52b62b453.
