--- colossalai/gemini/memory_tracer/__init__.py
@@ -1,5 +1,0 @@
-from .model_data_memtracer import GLOBAL_MODEL_DATA_TRACER
-from .memory_monitor import AsyncMemoryMonitor, SyncCudaMemoryMonitor
-from .memstats_collector import MemStatsCollector
-
-__all__ = ['AsyncMemoryMonitor', 'SyncCudaMemoryMonitor', 'MemStatsCollector', 'GLOBAL_MODEL_DATA_TRACER']
