+++ docs/colossalai/colossalai.utils.memory_tracer.async_memtracer.rst
@@ -1,5 +1,5 @@
 colossalai.utils.memory\_tracer.async\_memtracer
 ================================================
 
-.. automodule:: colossalai.gemini.memory_tracer.async_memtracer
+.. automodule:: colossalai.utils.memory_tracer.async_memtracer
    :members:
