+++ docs/colossalai/colossalai.utils.memory_tracer.model_data_memtracer.rst
@@ -1,5 +1,5 @@
 colossalai.utils.memory\_tracer.model\_data\_memtracer
 ======================================================
 
-.. automodule:: colossalai.gemini.memory_tracer.model_data_memtracer
+.. automodule:: colossalai.utils.memory_tracer.model_data_memtracer
    :members:
