+++ docs/colossalai/colossalai.utils.memory_tracer.rst
@@ -1,13 +1,13 @@
 colossalai.utils.memory\_tracer
 ===============================
 
-.. automodule:: colossalai.gemini.memory_tracer
+.. automodule:: colossalai.utils.memory_tracer
    :members:
 
 
 .. toctree::
    :maxdepth: 2
 
-   colossalai.gemini.memory_tracer.async_memtracer
-   colossalai.gemini.memory_tracer.memstats_collector
-   colossalai.gemini.memory_tracer.model_data_memtracer
+   colossalai.utils.memory_tracer.async_memtracer
+   colossalai.utils.memory_tracer.memstats_collector
+   colossalai.utils.memory_tracer.model_data_memtracer
